requirements.txt
pycrypto
distro
